move the devices
