+++ system_tests/test_undulator_system.py
@@ -6,7 +6,7 @@ from dodal.devices.undulator import Undulator
 SIM_INSERTION_PREFIX = "SR03S"
 
 ID_GAP_LOOKUP_TABLE_PATH: str = (
-    "./tests/devices/unit_tests/test_beamline_undulator_to_gap_lookup_table.txt"
+    "./tests/devices/_data/test_beamline_undulator_to_gap_lookup_table.txt"
 )
