+++ tests/devices/_data/test_display.configuration
