+++ tests/devices/_data/test_jCameraManZoomLevels.xml
