+++ tests/devices/test_undulator.py
@@ -17,7 +17,7 @@ from dodal.devices.undulator import (
 )
 
 ID_GAP_LOOKUP_TABLE_PATH: str = (
-    "./tests/devices/unit_tests/test_beamline_undulator_to_gap_lookup_table.txt"
+    "./tests/devices/_data/test_beamline_undulator_to_gap_lookup_table.txt"
 )
